backend/requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
Flask
Flask-SQLAlchemy
mysqlclient
Werkzeug